+++ src/main/java/com/github/yunabraska/githubworkflow/completion/CodeCompletionService.java
@@ -55,9 +55,7 @@ public class CodeCompletionService extends CompletionContributor {
                     @NotNull final CompletionResultSet resultSet
             ) {
                 getWorkflowFile(parameters.getPosition()).map(path -> yamlOf(parameters.getPosition())).map(YamlElement::context).ifPresent(context -> {
-                    // Fixme: remove if EDT errors pops up
                     final Project project = Optional.of(parameters.getOriginalFile()).map(PsiElement::getProject).orElse(null);
-                    context.actions().values().forEach(action -> action.resolve(parameters.getEditor().getProject()));
                     final int offset = parameters.getOffset();
                     final YamlElement position = context.getClosestElement(offset).orElse(new YamlElement(-1, -1, null, null, null, null));
                     final String[] prefix = new String[]{""};